+++ SQLiteStudio3/coreSQLiteStudio/plugins/dbpluginsqlite3.cpp
@@ -10,6 +10,9 @@ Db* DbPluginSqlite3::getInstance(const QString& name, const QString& path, const
 
     if (!db->openForProbing())
     {
+        if (errorMessage)
+            *errorMessage = db->getErrorText();
+
         delete db;
         return nullptr;
     }
@@ -17,6 +20,9 @@ Db* DbPluginSqlite3::getInstance(const QString& name, const QString& path, const
     SqlQueryPtr results = db->exec("SELECT * FROM sqlite_master");
     if (results->isError())
     {
+        if (errorMessage)
+            *errorMessage = db->getErrorText();
+
         delete db;
         return nullptr;
     }